requirements.txt
attrs==19.3.0
rhasspy-asr==0.1.1
pocketsphinx==0.1.15
